Refactors SmallConfigurator
