+++ src/components/small-configurator/TopComponent.tsx
@@ -1,0 +1,8 @@
+const TopComponent = () => (
+  <p className='max-w-max self-end rounded-xl bg-white p-2 text-right text-sm italic'>
+    Scroll door de onderstaande tabel naar beneden om alle beschikbare items te
+    zien en selecteer de werkitems die je nodig heb.
+  </p>
+)
+
+export { TopComponent }
