+++ src/components/small-configurator/formatPrice.ts
@@ -1,0 +1,3 @@
+const formatPrice = (price: number): string => `€${price},-`
+
+export { formatPrice }
